chore: Separate into a different file the task run functions from main essh.go file
